Move Modal back to common and update logic accordingly
