+++ src/components/pages/Comparison/ComparisonCard.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import ModalComponent from "./Modal";
+import ModalComponent from "../../common/Modal";
 import { Divider, Skeleton } from "antd";
 //icons
 import weather from "../../../styles/icons/weather-48.png";
